added scrollTo() to mitigate what appears to be a bug in Chromium
After successful registration, the view gets much smaller because the questionnaire is removed and a status message displayed instead. At this time, the user would typically be viewing the bottom of the page. Although Chromium correctly removes stuff from the DOM, it still reserves space for the removed items and does not scroll to the top. Firefox acts correctly.