enable html tags in informational texts (according to angular xss security model). fixes #5.
The caption text in CaptionItems, MultipleChoiceItems, and SingleChoiceItems are now bound to innerHTML rather than passed via interpolation. This allows users to use such tags as <b> (which I want) and <div> (which is an acceptable by-product), but not <script> (and neither things like <div onclick="...">) (which would be dangerous).

The sanitization is still done by Angular. See https://angular.io/guide/security#xss for more details.